✨feat: OpenApi 지수 정보 조회 클라이언트 구현
1. swagger 스펙을 반영하여 IndexInformation.baseIndex 타입을 Float로 수정
2.IndexOpenApiClient.syncIndexInformation 예외처리 아직 미완성

+++ src/main/java/com/ssibssaggi/findex/domain/entity/index/IndexInformation.java
@@ -19,7 +19,7 @@ public class IndexInformation {
     private String indexName;
     private Integer employedItemsCount;
     private LocalDate basePointInTime;
-    private Integer baseIndex;
+    private Float baseIndex;
     private SourceType sourceType;
     private Boolean favorite;
     private Boolean enabled;
